pyproject.toml
[tool.poetry]
name = "chart-differ"
version = "0.1.0"
description = ""
authors = ["Germain GAU <germain.gau@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
deepdiff = "^7.0.1"
pyyaml = "^6.0.1"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
